refactor: GameMap이 수행하던 맵에 O,X 그리는 작업을 bridgaMapPainter 클래스로 분리, GameMap은 초기 완성된 게임 맵만 가지도록 리팩토링
