+++ src/model/GameMap.js
@@ -1,0 +1,17 @@
+class GameMap {
+  #bridgeGameMap;
+
+  setBridgeGameMap(gameMap) {
+    this.#bridgeGameMap = gameMap;
+  }
+
+  getGameMap() {
+    return [...this.#bridgeGameMap];
+  }
+
+  getMapLength() {
+    return this.#bridgeGameMap.length;
+  }
+}
+
+module.exports = GameMap;
